# web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Laplace.Endpoints.OpenAICompat | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/recipe/compile": {
      "post": {
        "tags": [
          "foundry"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeCompileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecipeCompileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequiredResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/synthesis/export": {
      "post": {
        "tags": [
          "foundry"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SynthesisExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesisExportResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "402": {
            "description": "Payment Required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRequiredResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingReceipt": {
        "required": [
          "quote_id",
          "amount_cents",
          "currency",
          "tenant",
          "service_id"
        ],
        "type": "object",
        "properties": {
          "quote_id": {
            "type": "string"
          },
          "amount_cents": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "currency": {
            "type": "string"
          },
          "tenant": {
            "type": "string"
          },
          "service_id": {
            "type": "string"
          }
        }
      },
...
      "ErrorBody": {
        "required": [
          "type",
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "PaymentRequiredBody": {
        "required": [
          "type",
          "code",
          "message",
          "detail"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "detail": { }
        }
      },
      "PaymentRequiredResponse": {
        "required": [
          "error"
        ],
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/PaymentRequiredBody"
          }
        }
      },
...
      "RecipeCompileRequest": {
        "required": [
          "recipe"
        ],
        "type": "object",
        "properties": {
          "recipe": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RecipeCompileResponse": {
        "required": [
          "id",
          "object",
          "created",
          "recipe_id_hex",
          "name",
          "structure",
          "hidden_size",
          "num_layers",
          "compile_mode",
          "billing"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "created": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "recipe_id_hex": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "structure": {
            "type": "string"
          },
          "hidden_size": {
            "type": "string"
          },
          "num_layers": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "compile_mode": {
            "type": "string"
          },
          "billing": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BillingReceipt"
              }
            ]
          }
        }
      },
...
      "SynthesisExportRequest": {
        "required": [
          "recipe"
        ],
        "type": "object",
        "properties": {
          "recipe": {
            "type": [
              "null",
              "string"
            ]
          },
          "recipe_id_prefix": {
            "type": [
              "null",
              "string"
            ]
          },
          "tokenizer_dir": {
            "type": [
              "null",
              "string"
            ]
          },
          "format": {
            "type": [
              "null",
              "string"
            ]
          },
          "filename": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SynthesisExportResponse": {
        "required": [
          "id",
          "object",
          "created",
          "format",
          "output_path",
          "bytes",
          "billing"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "created": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "format": {
            "type": "string"
          },
          "output_path": {
            "type": "string"
          },
          "bytes": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int64"
          },
          "billing": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BillingReceipt"
              }
            ]
          }
        }
      },
...
    }
  },
...
}